refactor(ReservationTimeAvailability): ReservationTimeAvailability를 서비스 DTO로 이동
